+++ src/lmrtfy/__init__.py
@@ -5,12 +5,6 @@ import pathlib
 import logging
 
 
-#from lmrtfy.annotation import resource
-#from lmrtfy.annotation import variable
-#from lmrtfy.annotation import result
-#import lmrtfy.runner
-
-
 logging.basicConfig(format='%(asctime)s %(levelname)s: %(message)s', level=logging.DEBUG)
 
 # Don't import or use the underscored variables. These are subject to change.
@@ -31,3 +25,8 @@ if not _lmrtfy_profiles_dir.is_dir():
 
 if not _lmrtfy_config_dir.is_dir():
     os.mkdir(_lmrtfy_config_dir)
+
+from lmrtfy.annotation import resource
+from lmrtfy.annotation import variable
+from lmrtfy.annotation import result
+from lmrtfy import runner
